Generate (and export) simplified codebook

Generate (and export) simplified codebook

gensimplecodebook(
  survey_json,
  DKtext = "\\u308f\\u304b\\u3089\\u306a\\u3044",
  DKcode = 888,
  NAtext = "\\u7b54\\u3048\\u305f\\u304f\\u306a\\u3044",
  NAcode = 999,
  OtherOptiontext = "\\u305d\\u306e\\u4ed6",
  OtherOptioncode = 666,
  includeTimeStamp = TRUE,
  includeEmail = FALSE,
  text_randomize = NULL,
  text_starttime = NULL,
  show_duration = TRUE,
  textout = NULL,
  csvout = NULL
)

Arguments

survey_json

JSON survey file exported through surveyexport.js.

DKtext

Raw full text(s) of the response option(s) that indicate "don't know" answers.

DKcode

Response code for "don't know" answers.

NAtext

Raw full text(s) of the response option(s) that indicate refused answers.

NAcode

Response code for refused answers.

OtherOptiontext

Texts in the codebook to indicate "other" answers.

OtherOptioncode

Response code for "other" answers.

includeTimeStamp

Boolean to indicate whether to include system-generated timestamp variable in the codebook or not.

includeEmail

Boolean to indicate whether to include system-generated email variable in the codebook or not.

text_randomize

Raw full text(s) of the question(s) that are used to randomize experimental conditions.

text_starttime

Raw full text of the question that is used to record starting time.

show_duration

Boolean to indicate whether to include duration variable in the codebook or not (if text_starttime is available).

textout

Name and path of the output html file to be saved.

csvout

Name and path of the output csv file to be saved.